NoteToolbar Redux refactor
